requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cleanup==5.1.0
django-phonenumber-field==5.0.0
phonenumbers==8.12.16
Pillow==8.1.2
pytz==2020.5
sqlparse==0.4.1
